refactor(macos): consolidate close workflow
