+++ clients/apple/scripts/test-shell-runtime-metadata.sh
@@ -51,7 +51,7 @@ CLANG_MODULE_CACHE_PATH="$MODULE_CACHE_DIR" swiftc \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellControlFilePoller.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellActionCoordinator.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellAppActivityProvider.swift" \
-    "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellCloseConfirmationPresenter.swift" \
+    "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellCloseWorkflow.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellPasteboard.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellContentRenderingRegistry.swift" \
     "$REPO_ROOT/clients/apple/alan-macos/Services/Shell/ShellCoreFFIAdapter.swift" \
